0.1.0-Preview
This first realease provides a siple skill system that can be authored by Scriptable Object.
For now, it allow the definition of single target skill that are subject to a cast time and cooldown time.
Its allow to define any custom type of custom direct effect and comes with a Simple Skill sample taht demontrate how to defin a simple direct damage skill.